+++ src/Ev.ServiceBus/Dispatch/Outbox/OutboxMessageDispatcher.cs
@@ -6,8 +6,15 @@ namespace Ev.ServiceBus.Dispatch.Outbox;
 
 public class OutboxMessageDispatcher : IMessageDispatcher
 {
-    public Task ExecuteDispatches(CancellationToken token)
+    private readonly IOutboxService _service;
+
+    public OutboxMessageDispatcher(IOutboxService service)
     {
-        return Task.CompletedTask;
+        _service = service;
+    }
+
+    public async Task ExecuteDispatches(CancellationToken token)
+    {
+        await _service.EagerlySendStoredMessages(token);
     }
 }