+++ docs/CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 5.5.0
-- Added Outbox pattern support
+- Added Outbox pattern integration
+- Changed
+  - Optimized sending of messages by sending batches in parallel
 
 ## 5.4.0
 - Added Isolation feature V2